setup running the app"
